# Optimize double nested maps