In a few places, we use `map[]map[]x`. Unless we need to be able to iterate over all the inner maps, this is much slower than using a struct as a key

See tests below:

```go
func BenchmarkMulti(b *testing.B) {
	b.Run("nested-outer", func(b *testing.B) {
		var m map[string]map[string]struct{}
		for n := 0; n < b.N; n++ {
			m = map[string]map[string]struct{}{}
			for outer := 0; outer < 100; outer++ {
				o := strconv.Itoa(outer)
				for inner := 0; inner < 1; inner++ {
					i := strconv.Itoa(outer)
					if _, f := m[o]; !f {
						m[o] = map[string]struct{}{}
					}
					m[o][i] = struct{}{}
				}
			}
		}
	})
	b.Run("nested-inner", func(b *testing.B) {
		var m map[string]map[string]struct{}
		for n := 0; n < b.N; n++ {
			m = map[string]map[string]struct{}{}
			for outer := 0; outer < 1; outer++ {
				o := strconv.Itoa(outer)
				for inner := 0; inner < 100; inner++ {
					i := strconv.Itoa(outer)
					if _, f := m[o]; !f {
						m[o] = map[string]struct{}{}
					}
					m[o][i] = struct{}{}
				}
			}
		}
	})
	b.Run("nested-both", func(b *testing.B) {
		var m map[string]map[string]struct{}
		for n := 0; n < b.N; n++ {
			m = map[string]map[string]struct{}{}
			for outer := 0; outer < 100; outer++ {
				o := strconv.Itoa(outer)
				for inner := 0; inner < 100; inner++ {
					i := strconv.Itoa(outer)
					if _, f := m[o]; !f {
						m[o] = map[string]struct{}{}
					}
					m[o][i] = struct{}{}
				}
			}
		}
	})
	b.Run("tuple-outer", func(b *testing.B) {
		var m map[tuple]struct{}
		for n := 0; n < b.N; n++ {
			m = map[tuple]struct{}{}
			for outer := 0; outer < 100; outer++ {
				o := strconv.Itoa(outer)
				for inner := 0; inner < 1; inner++ {
					i := strconv.Itoa(outer)
					m[tuple{o, i}] = struct{}{}
				}
			}
		}
	})
	b.Run("tuple-inner", func(b *testing.B) {
		var m map[tuple]struct{}
		for n := 0; n < b.N; n++ {
			m = map[tuple]struct{}{}
			for outer := 0; outer < 1; outer++ {
				o := strconv.Itoa(outer)
				for inner := 0; inner < 100; inner++ {
					i := strconv.Itoa(outer)
					m[tuple{o, i}] = struct{}{}
				}
			}
		}
	})
	b.Run("tuple-both", func(b *testing.B) {
		var m map[tuple]struct{}
		for n := 0; n < b.N; n++ {
			m = map[tuple]struct{}{}
			for outer := 0; outer < 100; outer++ {
				o := strconv.Itoa(outer)
				for inner := 0; inner < 100; inner++ {
					i := strconv.Itoa(outer)
					m[tuple{o, i}] = struct{}{}
				}
			}
		}
	})
}

type tuple struct {
	a, b string
}
```

```
BenchmarkMulti
BenchmarkMulti/nested-outer
BenchmarkMulti/nested-outer-8              54590             22319 ns/op           27273 B/op        209 allocs/op
BenchmarkMulti/nested-inner
BenchmarkMulti/nested-inner-8             740229              1724 ns/op             448 B/op          4 allocs/op
BenchmarkMulti/nested-both
BenchmarkMulti/nested-both-8                4005            300571 ns/op           27282 B/op        209 allocs/op
BenchmarkMulti/tuple-outer
BenchmarkMulti/tuple-outer-8              120422              9589 ns/op           10507 B/op         11 allocs/op
BenchmarkMulti/tuple-inner
BenchmarkMulti/tuple-inner-8              631408              2026 ns/op             336 B/op          2 allocs/op
BenchmarkMulti/tuple-both
BenchmarkMulti/tuple-both-8                 5490            234907 ns/op           10516 B/op         11 allocs/op
```